Use composite indexes for selectivity estimation
* Pass PlanContext and aggregating properties to more places.
  They are used by the CompositeExpressionSelectivityCalculator
  to obtain index matches.

+++ community/cypher/front-end/util/src/main/scala/org/neo4j/cypher/internal/util/Selectivity.scala
@@ -21,7 +21,7 @@ import scala.language.implicitConversions
 case class Selectivity private(factor: Double) extends Ordered[Selectivity] {
   assert(factor >= 0 && factor <= 1.0, s"selectivity was $factor")
   def *(other: Selectivity): Selectivity = Selectivity(other.factor * factor)
-  def ^(a: Int): Selectivity = Selectivity(Math.pow(factor, a))
+  def ^(a: Double): Selectivity = Selectivity(Math.pow(factor, a))
   def negate: Selectivity = {
     val f = 1.0 - factor
     if (factor == 0 || f < 1)
